pyproject.toml
[tool.poetry]
name = "qwenbotq"
version = "0.1.0"
description = "A smart QQ bot based on Nonebot2 and Onebot v11."
authors = ["originalFactor"]
license = "GPL-v3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
nonebot2 = "^2.3.3"
nonebot-adapter-onebot = "^2.4.5"
websockets = "^13.0.1"
openai = "^1.47.0"
httpx = "^0.27.2"
httptools = "^0.6.1"
uvicorn = "^0.30.6"
motor = "^3.6.0"
beanie = "^1.27.0"

[tool.poetry.group.dev.dependencies]
nb-cli = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
